removed unique constraints, rebuilt endpoints to compensate
