+++ Readme.md
@@ -106,7 +106,6 @@ http://127.0.0.1:8000/login/
 
 http://127.0.0.1:8000/updateuser/
 {
-	"newUsername": "thisguy",
 	"newEmail": "NOPE!323"
 }
